Expose the hostname to python checks.
We move the hostname detection to the util package since it is needed by the agent, the python checks and soon dogstatsd.
